create scraper class and move file to lib
